LibCompat: rewrite `OSProcess` to use LibUnix's `#spawn:...`
...rather than Mariano's `OSSubprocess` package. The problem is that
`OSSubProcess` reaping logic assumes that all processes are spawned by
it so it may (does) reap processes spawned by LibUnix.

This commit makes `OSProcess` compatibility class to use `LibUnix`
to avoid this problem and throw an error is OSSubProcess is loaded.

A proper fix would be to change `OSSubprocess` to only reap processes
it has spawned (see [1])

[1] https://github.com/pharo-contributions/OSSubprocess/issues/76