LibUnix: pass down Pharo's environment when `posix_spawn()`ing children
This commit fixes a rather bad bug when environment of child processes
spawn by `spawn:stdin:stdout:stderr:exit:` was empty. While this may be
desirable sometimes, a useful default is to pass down Pharo's
environment (so chilren "inherit" the environment).

This commit does just that and it is not providing tools to pass down
custom environment. This will be added other commit, if needed.